requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.3
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==11.3.0
pkgconfig==1.5.5
pycparser==2.22
PyJWT==2.10.1
pytz==2025.2
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
